Move and update namespaces
